requirements.txt
google-genai
supabase
python-dotenv
streamlit